try to fix prepare connectivity problems

+++ package/cloudshell/cp/aws/domain/services/ec2/route_table.py
@@ -67,5 +67,5 @@ class RouteTablesService(object):
         :return:
         """
         for route in route_table.routes:
-            if route.state == 'blackhole':
+            if hasattr(route, 'state') and route.state == 'blackhole':
                 route.delete()
